+++ larky/src/main/java/com/verygood/security/larky/modules/globals/PythonBuiltins.java
@@ -393,64 +393,6 @@ public final class PythonBuiltins {
       defaultValue == Starlark.UNBOUND ? null : defaultValue);
   }
 
-  //override built-in type
-  @StarlarkMethod(
-    name = "type",
-    doc =
-      "Returns the type name of its argument. This is useful for debugging and "
-        + "type-checking. Examples:"
-        + "<pre class=\"language-python\">"
-        + "type(2) == \"int\"\n"
-        + "type([1]) == \"list\"\n"
-        + "type(struct(a = 2)) == \"struct\""
-        + "</pre>"
-        + "This function might change in the future. To write Python-compatible code and "
-        + "be future-proof, use it only to compare return values: "
-        + "<pre class=\"language-python\">"
-        + "if type(x) == type([]):  # if x is a list"
-        + "</pre>" +
-        "\n" +
-        "Type can overridden on any LarkyObject by implementing a __type__ special method." +
-        "Otherwise, the type will default to the default Starlark::type() method invocation",
-    parameters = {
-      @Param(name = "x", doc = "The object to check type of."),
-      @Param(name = "bases", defaultValue = "None"),
-      @Param(name = "dict", defaultValue = "None")
-    },
-    extraKeywords = @Param(name = "kwargs", defaultValue = "{}"),
-    useStarlarkThread = true
-  )
-  public Object type(Object object, Object bases, Object dict, Dict<String, Object> kwargs, StarlarkThread thread) throws EvalException {
-    if (Starlark.isNullOrNone(bases) && Starlark.isNullOrNone(dict) && kwargs.size() == 0) {
-      // There is no 'type' type in Starlark, so we return a string with the type name.
-      if (LarkyObject.class.isAssignableFrom(object.getClass())) {
-        return ((LarkyObject) object).type();
-      }
-      return Starlark.type(object);
-    } else if (kwargs.size() != 0) {
-      throw Starlark.errorf("type() takes 1 or 3 arguments");
-    }
-    return Starlark.type(object); // TODO: fix.
-      /*
-           Collection<String> fieldNames =
-          fields instanceof Sequence
-              ? Sequence.cast(fields, String.class, "fields")
-              : fields instanceof Dict
-              ? Dict.cast(fields, String.class, String.class, "fields").keySet()
-              : null;
-
-      if(!Strings.isNullOrEmpty(name)) {
-         return LarkyType.createExportedSchemaful(
-             new LarkyType.Key("BUILTIN", name),
-             fieldNames,
-             thread.getCallerLocation()
-         );
-      }
-      return LarkyType.createUnexportedSchemaful(fieldNames, thread.getCallerLocation());
-       */
-  }
-
-
   @StarlarkMethod(
     name = "hash",
     doc =
