+++ larky/src/test/resources/stdlib_tests/test_larky.star
@@ -13,7 +13,7 @@ def _test_namespace_exposes_larky_builtins():
     :return: None
     """
     items = sorted(dir(larky))
-    asserts.assert_that(items).is_length(16)
+    asserts.assert_that(items).is_length(17)
     asserts.assert_that(items).is_equal_to(sorted([
         "SENTINEL",
         "mutablestruct",
@@ -31,6 +31,7 @@ def _test_namespace_exposes_larky_builtins():
         "impl_function_name",
         "DeterministicGenerator",
         "dicts",
+        "is_subclass",
     ]))
